An Example of a GAME MODE with various utilities
This Game Mode class has a few other actor classes related and spawns one of these, from a BP class or from a C++ class. It will also bind a delegate to these classes in case they want to answer to a generic concept I created as an example. 
The YOURGAMEManager class makes a generic registry of all the actors that are spawned from the CustomActor class.